% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_class.R
\name{change_class}
\alias{change_class}
\title{Change Column Classes in a Data Frame}
\usage{
change_class(
  data,
  to_numeric = NULL,
  to_integer = NULL,
  to_character = NULL,
  to_factor = NULL,
  to_logical = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} that contains the variables whose types are to be changed.}

\item{to_numeric}{Column indexes (or names) of variables in \code{data} to be converted to \code{numeric}.}

\item{to_integer}{Column indexes (or names) of variables in \code{data} to be converted to \code{integer}.}

\item{to_character}{Column indexes (or names) of variables in \code{data} to be converted to \code{character}.}

\item{to_factor}{Column indexes (or names) of variables in \code{data} to be converted to \code{factor}.}

\item{to_logical}{Column indexes (or names) of variables in \code{data} to be converted to \code{logical}.}
}
\value{
A \code{data.frame} with the specified columns converted to the desired class types.
}
\description{
This function changes the class of specified columns in a data frame to one of several types: numeric, integer, character, factor, or logical. By default, column indexes are expected. If column names are provided, they will be converted to their corresponding indexes.
}
\details{
The function converts the specified columns to the desired class. By default, column indexes should be provided as integers. If column names are passed, they will be automatically converted to their corresponding column indexes. The function will stop if there are duplicated column references across the different conversion categories.
}
