% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{cobb_douglas_CRS_XnY1}
\alias{cobb_douglas_CRS_XnY1}
\title{1 output ~ nX inputs Cobb-Douglas CRS Data Generation Process}
\usage{
cobb_douglas_CRS_XnY1(N, nX)
}
\arguments{
\item{N}{Sample size.}

\item{nX}{Number of inputs. Possible values: \code{1}, \code{3}, \code{6}, \code{9}, \code{12} and \code{15}.}
}
\value{
\code{data.frame} with the simulated data: nX inputs, 1 output (y) and the theoretical frontier (yD).
}
\description{
This function is used to simulate a 1 output ~ nX inputs \code{data.frame} with a Cobb-Douglas CRS Data Generation Process.
}
\keyword{internal}
