% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{cobb_douglas_X3Y3}
\alias{cobb_douglas_X3Y3}
\title{3 outputs ~ 3 inputs Cobb-Douglas Data Generation Process}
\usage{
cobb_douglas_X3Y3(N, border, noise, returns)
}
\arguments{
\item{N}{Sample size.}

\item{border}{Proportion of DMUs in the frontier: \code{0.1}, \code{0.2} or \code{0.3}.}

\item{noise}{Random noise. 4 possible levels: \code{0}, \code{0.5}, \code{1} or \code{2}.}

\item{returns}{Returns to scale. \code{"CRS"} for Constant Return to Scale, \code{"DRS"} for Decreasing Return to Scale and \code{"IRS"} for Increasing Return to Scale.}
}
\value{
\code{data.frame} with the simulated data: 3 inputs (x1, x2, x3), 3 outputs (y1, y2, y3) and the theoretical frontier (yD1, yD2, yD3)
}
\description{
This function is used to simulate data a 3 outputs ~ 3 inputs \code{data.frame}  with a Cobb-Douglas Data Generation Process as in \insertCite{ahn2023;textual}{aces}.
}
\references{
\insertRef{ahn2023}{aces}
}
\keyword{internal}
