% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_scores_additive}
\alias{compute_scores_additive}
\title{Data envelopment analysis' clasification.}
\usage{
compute_scores_additive(data, x, y)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}
}
\value{
Fill
}
\description{
This function trains for each model, the different hyperparameters and returns the best model with its best hyperparameters.
}
