% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probabilities.R
\name{compute_target}
\alias{compute_target}
\title{Projections to the Hyperplane}
\usage{
compute_target(
  data,
  x,
  y,
  z = NULL,
  final_model,
  orientation,
  cut_off,
  imp_vector
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of environment variables in \code{data}.}

\item{final_model}{The best-fitted model used for efficiency score computation.}

\item{orientation}{The direction in which data should be projected to calculate efficiency scores.}

\item{cut_off}{Probability levels for determining efficient class scores.}

\item{imp_vector}{A \code{data.frame} with importance variables results}
}
\value{
A numeric vector containing the efficiency scores for each observation in the input data.
}
\description{
This function computes the efficiency scores based on a given model.
}
