% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_scores.R
\name{ddf}
\alias{ddf}
\title{The Directional Distance Function}
\usage{
ddf(tech_xmat, tech_ymat, eval_xmat, eval_ymat, direction, convexity, returns)
}
\arguments{
\item{tech_xmat}{A \code{data.frame} or \code{matrix} containing the observed inputs to determine the technology.}

\item{tech_ymat}{A \code{data.frame} or \code{matrix} containing the observed outputs to determine the technology.}

\item{eval_xmat}{A \code{data.frame} or \code{matrix} containing the containing the input data of the DMUs to be evaluated.}

\item{eval_ymat}{A \code{data.frame} or \code{matrix} containing the containing the output data of the DMUs to be evaluated.}

\item{direction}{Direction of the vector to project on the frontier. Two possibilities:
\itemize{
\item{\code{"mean"}} Projection vector given by the average value of inputs and outputs of all DMUs.
\item{\code{"briec"}} Projection vector given by the value of inputs and outputs of the evaluated DMU.
}}

\item{convexity}{A \code{logical} value indicating if a convex technology is assumed.}

\item{returns}{Type of returns to scale.}
}
\value{
A \code{vector} of \code{"numeric"} scores computed through the input-oriented radial model in the envelopment format.
}
\description{
This function computes efficiency scores through the directional distance function in the envelopment format.
}
