% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dea.R
\name{is.dea}
\alias{is.dea}
\title{dea class check.}
\usage{
is.dea(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a \code{dea} object (that is, has "dea"
amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks whether an R object is of dea class or not.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
