% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.deadata.R
\name{is.deadata}
\alias{is.deadata}
\title{deadata class check.}
\usage{
is.deadata(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a \code{deadata} object (that is, has "deadata"
amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks whether an R object is of deadata class or not.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
