% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{mult_scenario_XnY1}
\alias{mult_scenario_XnY1}
\title{1 output ~ nX inputs Multiplicative Data Generation Process}
\usage{
mult_scenario_XnY1(N, scenario)
}
\arguments{
\item{N}{Sample size.}

\item{scenario}{\code{"A"}, \code{"B"}, \code{"C"}, \code{"D"}, \code{"E"} or \code{"F"}. For details, check Table 2 of \insertCite{kuosmanen2010;textual}{aces}.}
}
\value{
\code{data.frame} with the simulated data: 1-3 inputs, 1 output (y) and the theoretical frontier (yD).
}
\description{
This function is used to simulate a 1 output ~ nX inputs \code{data.frame} with a Multiplicative Data Generation Process as in \insertCite{kuosmanen2010;textual}{aces}.
}
\references{
\insertRef{kuosmanen2010}{aces}
}
\keyword{internal}
