% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAX_labeled.R
\name{PEAX_labeled}
\alias{PEAX_labeled}
\title{Labeled and Treatment Database for Probabilistic Efficiency Analysis with Explicable Artificial Intelligence (PEAX)}
\usage{
PEAX_labeled(
  data,
  x,
  y,
  z_c = NULL,
  z_d = NULL,
  valid_data = NULL,
  orientation_labeled = "out",
  p = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the original dataset used for training and labeling.}

\item{x}{Column indexes or names of input variables in \code{data}.}

\item{y}{Column indexes or names of output variables in \code{data}.}

\item{z_c}{Column indexes or names of continuous contextual variables (optional).}

\item{z_d}{Column indexes or names of discrete (categorical) contextual variables (optional).}

\item{valid_data}{A separate \code{data.frame} or \code{matrix} with validation data to assess generalization performance.}

\item{orientation_labeled}{A \code{character} string specifying the orientation of the model: either \code{"input"} or \code{"output"}.}

\item{p}{A \code{numeric} value or vector of values in (0,1), representing the quantile threshold(s) used to label efficient units. When multiple values are provided, the function generates one labeled dataset per threshold, allowing users to explore different levels of class imbalance in the efficiency classification.}
}
\value{
A labeled dataset with efficiency classes, contextual variables, and treatment indicators, ready for modeling in the PEAX framework.
}
\description{
This function prepares and labels a dataset for efficiency analysis under the Conditional DEA framework, as implemented in the PEAX methodology (Probabilistic Efficiency Analysis with Explicable Artificial Intelligence). It classifies Decision Making Units (DMUs) into efficient or not efficient units. The function explicitly addresses the problem of class imbalance—common in frontier estimation—by allowing the user to control the degree of imbalance. It supports both continuous and discrete environmental variables.
}
\details{
This function is part of the PEAX methodology (Probabilistic Efficiency Analysis with Explicable Artificial Intelligence). It classifies DMUs based on a quantile frontier and supports environmental heterogeneity through contextual variables. Labels generated can be used in classification models to approximate the efficiency frontier probabilistically.
}
