% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance_glm}
\alias{performance_glm}
\title{Training a Classification Machine Learning Model}
\usage{
performance_glm(model_fit, new_data)
}
\arguments{
\item{model_fit}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{new_data}{Parameters for controlling the training process (from the \code{'caret'} package).}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains a set of models and selects best hyperparameters for each of them.
}
