% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Prepare Data and Handle Errors}
\usage{
preprocessing(data, x, y)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}
}
\value{
It returns a \code{matrix} in the required format and displays some error messages.
}
\description{
This function arranges the data in the required format and displays some error messages.
}
