% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_estimation.R
\name{select_cut_off}
\alias{select_cut_off}
\title{Select cut-off point in Classification Algorithm}
\usage{
select_cut_off(data, final_model)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{final_model}{The best \code{train} object from \code{caret}.}
}
\value{
It returns the best cut-off point.
}
\description{
This function selects the cut-off point that minimizes the sum of false positives and false negatives.
}
