% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE_data.R
\name{SMOTE_data}
\alias{SMOTE_data}
\title{Create New SMOTE Units to Balance Data combinations of m + s}
\usage{
SMOTE_data(data, x, y, RTS = "vrs", balance_data, seed, verbose = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} containing the variables used in the model.}

\item{x}{Column indexes of the input variables in the \code{data}.}

\item{y}{Column indexes of the output variables in the \code{data}.}

\item{RTS}{Text string or number defining the underlying DEA technology /
returns-to-scale assumption (default: \code{"vrs"}). Accepted values:
\describe{
\item{\code{1} / \code{"vrs"}}{Variable returns to scale, convexity and free disposability.}
\item{\code{3} / \code{"crs"}}{Constant returns to scale, convexity and free disposability.}
}}

\item{balance_data}{Indicate level of efficient units to achive and the number of efficient and not efficient units.}

\item{seed}{Integer. Seed for reproducibility.}

\item{verbose}{Logical; if \code{TRUE}, prints progress messages (default \code{FALSE}).}
}
\value{
It returns a \code{data.frame} with the newly created set of DMUs incorporated.
}
\description{
This function creates new DMUs to address data imbalances. If the majority class is efficient, it generates new inefficient DMUs by worsering the observed units. Conversely, if the majority class is inefficient, it projects inefficient DMUs to the frontier. Finally, a random selection if performed to keep a proportion of 0.65 for the majority class and 0.35 for the minority class.
}
\keyword{internal}
