% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_glm}
\alias{train_glm}
\title{Training a Classification Machine Learning Model}
\usage{
train_glm(data, method, arguments)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{method}{Parameters for controlling the training process (from the \code{'caret'} package).}

\item{arguments}{A \code{list} of selected machine learning models and their hyperparameters.}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains a set of models and selects best hyperparameters for each of them.
}
