% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_PEAXAI}
\alias{train_PEAXAI}
\title{Training a Classification Machine Learning Model}
\usage{
train_PEAXAI(data, method, parameters, trControl, metric_priority)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{method}{Parameters for controlling the training process (from the \code{'caret'} package).}

\item{parameters}{A \code{list} of selected machine learning models and their hyperparameters.}

\item{trControl}{A \code{list} of selected machine learning learning.}

\item{metric_priority}{dfdfd}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains a set of models and selects best hyperparameters for each of them.
}
