% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{translog_X2Y2}
\alias{translog_X2Y2}
\title{2 outputs ~ 2 inputs Translog Data Generation Process}
\usage{
translog_X2Y2(N, border, noise)
}
\arguments{
\item{N}{Sample size.}

\item{border}{Proportion of DMUs in the frontier.}

\item{noise}{\code{logical}. Random noise.}
}
\value{
\code{data.frame} with the simulated data: 2 inputs (x1, x2), 2 outputs (y1, y2) and the theoretical frontier (yD1, yD2).
}
\description{
This function is used to simulate a 2 outputs ~ 2 inputs \code{data.frame} with a Translog Data Generation Process as in \insertCite{santin2009;textual}{aces}.
}
\references{
\insertRef{santin2009}{aces}
}
\keyword{internal}
