% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{validate_fitting_args}
\alias{validate_fitting_args}
\title{Handle Errors on PEAXAI_fitting}
\usage{
validate_fitting_args(
  data,
  x,
  y,
  RTS,
  trControl,
  methods,
  metric,
  importance_method,
  hold_out,
  balance_data,
  scenarios
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}
}
\value{
It returns a \code{matrix} in the required format and displays some error messages.
}
\description{
This function check if PEAXAI arguments are well introduced and check if the analysis it will be done.
}
