% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAXAI_targets.R
\name{find_beta_maxmin}
\alias{find_beta_maxmin}
\title{Projections to the Hyperplane}
\usage{
find_beta_maxmin(
  data,
  x,
  y,
  final_model,
  efficiency_thresholds,
  n_expand,
  vector_gx,
  vector_gy,
  max_y,
  min_x
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{final_model}{The best-fitted model used for efficiency score computation.}

\item{efficiency_thresholds}{Probability levels for determining efficient class scores.}

\item{n_expand}{n_expand}

\item{vector_gx}{dfd}

\item{vector_gy}{dfdfd}

\item{max_y}{fd max_y}

\item{min_x}{fd min_x}

\item{directional_vector}{A \code{data.frame} with importance variables results}
}
\value{
A numeric vector containing the efficiency scores for each observation in the input data.
}
\description{
This function computes the efficiency scores based on a given model.
}
