% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAXAI_importance_features.R
\name{PEAXAI_global_importance}
\alias{PEAXAI_global_importance}
\title{Training Classification Models to Estimate Efficiency}
\usage{
PEAXAI_global_importance(
  data,
  x,
  y,
  final_model,
  background = "train",
  target = "train",
  importance_method
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{final_model}{A model fitted to use.}

\item{background}{The data samples used to define the model’s background distribution. Use the training set ("train") to reflect the distribution the model learned from, or the real-world data ("real") to assess model robustness under the actual data domain.}

\item{target}{The data to be explained/evaluated by the model. Selecting "train" allows understanding how the model makes decisions, while selecting "real" focuses on analyzing the underlying nature of the problem.}

\item{importance_method}{A \code{data.frame} with the technique and parameters.}
}
\value{
A \code{"cafee"} object.
}
\description{
Trains one or multiple classification algorithms to identify Pareto-efficient
decision-making units (DMUs). It jointly searches model hyperparameters and the
class-balancing level (e.g., synthetic samples via SMOTE) using k-fold cross-
validation or a train/validation/test split, selecting the configuration that
maximizes the specified metric(s). Returns, for each technique, the best fitted
model together with training summaries, performance metrics, and the selected
balancing level.
}
