% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAXAI_targets.R
\name{PEAXAI_targets}
\alias{PEAXAI_targets}
\title{Projections to the Hyperplane}
\usage{
PEAXAI_targets(
  data,
  x,
  y,
  final_model,
  efficiency_thresholds,
  directional_vector,
  n_expand,
  n_grid,
  max_y = 2,
  min_x = 1
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{final_model}{The best-fitted model used for efficiency score computation.}

\item{efficiency_thresholds}{Probability levels for determining efficient class scores.}

\item{directional_vector}{A \code{data.frame} with importance variables results}

\item{n_expand}{n_expand}

\item{n_grid}{n_grid}

\item{max_y}{fd max_y}

\item{min_x}{fd min_x}

\item{z}{Column indexes of environment variables in \code{data}.}
}
\value{
A numeric vector containing the efficiency scores for each observation in the input data.
}
\description{
This function computes the efficiency scores based on a given model.
}
