% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{reffcy}
\alias{reffcy}
\title{Random Sample for Efficiency Analysis}
\usage{
reffcy(DGP, parms)
}
\arguments{
\item{DGP}{Data Generation Process:
\itemize{
\item{\code{"cobb_douglas_XnY1"}}: Cobb-Douglas Data Generation Process for a XnY1 scenario. Check \code{help("cobb_douglas_XnY1")}.
\item{\code{"cobb_douglas_XnY1_CRS"}}: Cobb-Douglas Data Generation Process with CRS for a XnY1 scenario. Check \code{help("cobb_douglas_XnY1_CRS")}.
\item{\code{"translog_X2Y2"}}: Translog Data Generation Process for a X2Y2 scenario. Check \code{help("translog_X2Y2")}.
\item{\code{"add_scenario_XnY1"}}: Additive Data Generation Process for a XnY1 scenario. Check \code{help("add_scenario_XnY1")}.
\item{\code{"mult_scenario_XnY1"}}: Multiplicative Data Generation Process for a XnY1 scenario. Check \code{help("mult_scenario_XnY1")}.
\item{\code{"cobb_douglas_X3Y3"}}: Cobb-Douglas Data Generation Process for a X3Y3 scenario. Check \code{help("cobb_douglas_X3Y3")}.
}}

\item{parms}{\code{list} with the parameters of the simulation functions.
\itemize{
\item{\code{"N"}}: Sample size.
\item{\code{"nX"}}: Number of inputs. Available for the \code{cobb_douglas_XnY1(N, nX)} function.
\item{\code{"border"}}: Proportion of DMUs in the frontier. Available for the \code{translog_X2Y2(N, border, noise)} and the \code{cobb_douglas_X3Y3(N, border, noise, returns)} functions.
\item{\code{"noise"}}: \code{logical} indicating presence of random noise. Available for the \code{translog_X2Y2(N, border, noise)} and the \code{cobb_douglas_X3Y3(N, border, noise, returns)} functions.
\item{\code{"scenario"}}: Determine the type of Data Generation Process. Available for the \code{add_scenario_XnY1(N, scenario)} and the \code{mult_scenario_XnY1(N, scenario)} functions.
\item{\code{"returns"}}: Returns to scale. Available for the \code{cobb_douglas_X3Y3(N, border, noise, returns)} function.
}}
}
\value{
\code{data.frame} simulated with the selected Data Generation Process.
}
\description{
This function is used to simulate a \code{data.frame} with some given Data Generation Process.
}
\details{
Please refer to the help manuals of the mentioned functions for usage examples and more detailed information on the parameters.
}
