% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xai_prepare_sets.R
\name{xai_prepare_sets}
\alias{xai_prepare_sets}
\title{Training a Classification Machine Learning Model}
\usage{
xai_prepare_sets(
  data,
  x,
  y,
  final_model,
  background,
  target,
  type,
  threshold,
  levels_order
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{method}{Parameters for controlling the training process (from the \code{'caret'} package).}

\item{arguments}{A \code{list} of selected machine learning models and their hyperparameters.}

\item{hyparameter}{A \code{list}}

\item{metric}{A \code{string} specifying the summary metric for classification to select the optimal model. Default includes \code{"Balanced_accuracy"} due to (normally) unbalanced data.}
}
\value{
It returns a \code{list} with the chosen model.
}
\description{
This function trains a set of models and selects best hyperparameters for each of them.
}
